# fd_readdir: reverts to special cased dirent cache