@codefromthecrypt 2023-07-04
whoops we are missing a test that's in wasi-testsuite. I'll backfill

```
 right: `3`: expected three entries', src/bin/fd_readdir.rs:143:5
```