@codefromthecrypt 2023-07-04
re-ran the benchmark because the last commit, needed to pass wasi-testsuite, costs a little more memory. However, we're basically at the performance we had prior to refactoring, plus much better if using `fs.FS`.